Renamed task views and view models (#1987)
* Renamed "Task" to "VolunteerTask". Adjusted the namespaces as well, which cascaded to quite a few classes. Sorted out the namespace includes as well, based on the established convention.

* Renamed Testing namespace and test class for ViewModels\Validators\VolunteerTask\* to match the rename from Task to VolunteerTask.
