+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Tasks/DeleteQueryHandlerShould.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.Threading.Tasks;
+
 using AllReady.Areas.Admin.Features.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Models;
+
 using Xunit;
 
 namespace AllReady.UnitTest.Areas.Admin.Features.Tasks
