+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Tasks/EditVolunteerTaskCommandHandlerShould.cs
@@ -1,13 +1,15 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
-using AllReady.Areas.Admin.Features.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
-using AllReady.Models;
-using Xunit;
 using System.Linq;
-using Moq;
+using System.Threading.Tasks;
+
+using AllReady.Areas.Admin.Features.Tasks;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
+using AllReady.Models;
 using AllReady.Services;
+
+using Xunit;
+using Moq;
 
 namespace AllReady.UnitTest.Areas.Admin.Features.Tasks
 {
