+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Features/Tasks/MessageTaskVolunteersCommandHandlerShould.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+
 using AllReady.Areas.Admin.Features.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Features.Notifications;
 using AllReady.Models;
+
 using MediatR;
 using Moq;
 using Xunit;
