+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Controllers/TaskController.cs
@@ -1,10 +1,11 @@
-﻿using System.Collections.Generic;
+﻿using AllReady.Areas.Admin.Features.Tasks;
+using AllReady.Areas.Admin.ViewModels.Validators.VolunteerTask;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
+
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using AllReady.Areas.Admin.Features.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
-using AllReady.Areas.Admin.ViewModels.Validators.Task;
-using AllReady.Security;
+
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
