+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Features/Tasks/DetailsQuery.cs
@@ -1,4 +1,5 @@
-﻿using AllReady.Areas.Admin.ViewModels.Task;
+﻿using AllReady.Areas.Admin.ViewModels.VolunteerTask;
+
 using MediatR;
 
 namespace AllReady.Areas.Admin.Features.Tasks
