+++ AllReadyApp/Web-App/AllReady/Areas/Admin/ViewModels/Event/EventDetailViewModel.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+
 using AllReady.Areas.Admin.ViewModels.Itinerary;
 using AllReady.Areas.Admin.ViewModels.Shared;
-using AllReady.Areas.Admin.ViewModels.Task;
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
 using AllReady.Models;
 
 namespace AllReady.Areas.Admin.ViewModels.Event
