+++ AllReadyApp/Web-App/AllReady/Areas/Admin/ViewModels/Validators/VolunteerTask/IValidateVolunteerTaskEditViewModelValidator.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
-using AllReady.Areas.Admin.ViewModels.Task;
 
-namespace AllReady.Areas.Admin.ViewModels.Validators.Task
+using AllReady.Areas.Admin.ViewModels.VolunteerTask;
+
+namespace AllReady.Areas.Admin.ViewModels.Validators.VolunteerTask
 {
     public interface IValidateVolunteerTaskEditViewModelValidator
     {
