+++ AllReadyApp/Web-App/AllReady/Areas/Admin/ViewModels/VolunteerTask/MessageTaskVolunteersViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace AllReady.Areas.Admin.ViewModels.Task
+namespace AllReady.Areas.Admin.ViewModels.VolunteerTask
 {
     public class MessageTaskVolunteersViewModel
     {
