+++ AllReadyApp/Web-App/AllReady/Areas/Admin/ViewModels/VolunteerTask/TaskSummaryViewModel.cs
@@ -1,12 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.Linq;
+
 using AllReady.Areas.Admin.ViewModels.Shared;
 using AllReady.Models;
 
-namespace AllReady.Areas.Admin.ViewModels.Task
-{
-    public class DetailsViewModel
+namespace AllReady.Areas.Admin.ViewModels.VolunteerTask
+{  
+    public class TaskSummaryViewModel
     {
         public int Id { get; set; }
 
@@ -15,10 +17,16 @@ namespace AllReady.Areas.Admin.ViewModels.Task
         [Display(Name = "Event")]
         public string EventName { get; set; }
 
+        public DateTimeOffset EventStartDateTime { get; set; }
+
+        public DateTimeOffset EventEndDateTime { set; get; }
+
         public int CampaignId { get; set; }
 
         [Display(Name = "Campaign")]
         public string CampaignName { get; set; }
+
+        public int OrganizationId { get; set; }
 
         [Required]
         public string Name { get; set; }
@@ -29,7 +37,7 @@ namespace AllReady.Areas.Admin.ViewModels.Task
 
         [Display(Name = "Start date")]
         public DateTimeOffset StartDateTime { get; set; }
-
+        
         [Display(Name = "End date")]
         public DateTimeOffset EndDateTime { get; set; }
 
@@ -43,7 +51,8 @@ namespace AllReady.Areas.Admin.ViewModels.Task
 
         public List<VolunteerTaskSkill> RequiredSkills { get; set; } = new List<VolunteerTaskSkill>();
 
-        /// <summary>List of file attachments</summary>
-        public List<FileAttachment> Attachments { get; set; } = new List<FileAttachment>();
+        public int AcceptedVolunteerCount => this.AssignedVolunteers?.Count(v => v.HasVolunteered) ?? 0;
+
+        public int NumberOfVolunteersSignedUp => this.AssignedVolunteers.Count;
     }
 }
