--- AllReadyApp/Web-App/AllReady/Areas/Admin/Views/Task/Details.cshtml
@@ -1,151 +1,0 @@
-﻿@using System.Threading.Tasks
-@using AllReady.Areas.Admin.ViewModels.Task
-@using AllReady.Security
-@model DetailsViewModel
-@{
-    ViewData["Title"] = "Tasks Details";
-    var userTimeZoneId = User.GetTimeZoneId();
-}
-<div class="row">
-    <div class="col-12">
-        <ol class="breadcrumb">
-            <li><a asp-controller="Campaign" asp-action="Index" asp-area="Admin">Campaigns</a></li>
-            <li><a asp-controller="Campaign" asp-action="Details" asp-route-id="@Model.CampaignId" asp-area="Admin">@Model.CampaignName</a></li>
-            <li><a asp-controller="Event" asp-action="Details" asp-route-id="@Model.EventId" asp-area="Admin">@Model.EventName</a></li>
-            <li>@Model.Name</li>
-        </ol>
-    </div>
-</div>
-
-<div class="row">
-    <div class="col-md-10">
-        <h2>@Model.Name <a asp-area="Admin" asp-controller="Task" asp-action="Edit" asp-route-id="@Model.Id" class="btn btn-default"><i class="fa fa-edit"></i></a> <a asp-area="Admin" asp-controller="Task" asp-action="Delete" asp-route-id="@Model.Id" class="btn btn-danger"><i class="glyphicon glyphicon-trash"></i></a></h2>
-        <p>@Model.Description</p>
-        <p>
-            <time value="Model.StartDateTime"></time> - <time value="Model.EndDateTime"></time>
-            <em><time-zone-name time-zone-id="@Model.TimeZoneId"></time-zone-name></em>
-        </p>
-        @if (!string.IsNullOrEmpty(userTimeZoneId) && Model.TimeZoneId != userTimeZoneId)
-        {
-            <p>
-                <time value="Model.StartDateTime" target-time-zone-id="@userTimeZoneId"></time> - <time value="Model.EndDateTime" target-time-zone-id="@userTimeZoneId"></time>
-                <em><time-zone-name time-zone-id="@userTimeZoneId"></time-zone-name></em>
-            </p>
-        }
-
-        @if (Model.RequiredSkills.Any())
-        {
-            <div class="row">
-                <div class="col-md-12">
-                    <h3>Required Skills</h3>
-                    <ul>
-                        @foreach (var requiredSkill in Model.RequiredSkills)
-                {
-                            <li><strong>@requiredSkill.Skill.HierarchicalName</strong></li>
-                        }
-                    </ul>
-                </div>
-            </div>
-        }
-
-        @if (Model.Attachments.Any())
-        {
-            <div class="row">
-                <div class="col-md-12">
-                    <h3>Attachments</h3>
-                    <ul>
-                        @foreach (var attachment in Model.Attachments)
-                        {
-                            <li><a href="@attachment.Url">@attachment.Name</a>: @attachment.Description</li>
-                        }
-                    </ul>
-                </div>
-            </div>
-        }
-
-        <h3>
-            Volunteers @if (Model.NumberOfVolunteersRequired > 0)
-            { <span class="text-muted">(@Model.NumberOfVolunteersRequired Required)</span> }
-        </h3>
-
-        <div class="row">
-            <div class="col-md-12">
-                @if (Model.AssignedVolunteers.Any())
-                {
-                    <button type="button" id="openMessageToVolunteers" class="btn btn-primary" disabled="disabled" autocomplete="off" data-toggle="modal" data-target="#messageVolunteersModal">Message All</button>
-                    <table class="table">
-                        <tr>
-                            <th>Volunteer Name</th>
-                            <th>Phone</th>
-                            <th>Skills</th>
-                            <th>E-mail</th>
-                        </tr>
-                        @foreach (var volunteer in Model.AssignedVolunteers)
-                        {
-                            <tr>
-                                <td>@volunteer.Name</td>
-                                <td>@volunteer.PhoneNumber</td>
-                                <td>
-                                    <ul>
-                                        @foreach (var associatedSkill in volunteer.AssociatedSkills)
-                                        {
-                                            <li><strong>@associatedSkill.Skill.HierarchicalName</strong></li>
-                                        }
-                                    </ul>
-                                </td>
-                                <td>@volunteer.UserName</td>
-                            </tr>
-                        }
-                    </table>
-                }
-                else
-                {
-                    <div>No volunteers yet.</div>
-                }
-            </div>
-        </div>
-    </div>
-</div>
-<!-- Modal -->
-<div class="modal fade" id="messageVolunteersModal" tabindex="-1" role="dialog" aria-labelledby="messageVolunteersLabel">
-    <div class="modal-dialog" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                <h4 class="modal-title" id="messageVolunteersLabel">Message All Volunteers</h4>
-            </div>
-            <form asp-area="Admin" asp-controller="Task" asp-action="MessageAllVolunteers" method="post">
-                <div class="modal-body">
-                    <input type="hidden" value="@Model.Id" name="volunteerTaskId" id="message-volunteerTaskId">
-                    <div class="form-group">
-                        <label for="subject" class="control-label">Subject</label>
-                        <input type="text" class="form-control wide" name="subject" id="messageVolunteersModal-subject">
-                    </div>
-                    <div class="form-group">
-                        <label for="message-text" class="control-label">Message</label>
-                        <textarea rows="4" class="form-control wide" name="message" id="messageVolunteersModal-message"></textarea>
-                        <span id="messageCharacterCount" class="text-muted"></span>
-                    </div>
-                    <div class="alert alert-info" role="alert">Sending message...</div>
-                    <div class="alert alert-success" role="alert"><i class="fa fa-check"></i> Message sent</div>
-                    <div class="alert alert-danger" role="alert"></div>
-
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                    <button type="submit" id="sendMessageToVolunteers" class="btn btn-primary">Send Message</button>
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
-
-@section scripts {
-    <script src="~/js/taskDetails.js"></script>
-    <script type="text/javascript">
-        $(function () {
-            var taskDetailsAdmin = new HTBox.TaskDetailAdmin();
-            $("#openMessageToVolunteers").prop('disabled', false);
-        });
-    </script>
-}
