--- AllReadyApp/Web-App/AllReady/Areas/Admin/Views/Task/_List.cshtml
@@ -1,57 +1,0 @@
-﻿@model IEnumerable<AllReady.Areas.Admin.ViewModels.Task.TaskSummaryViewModel>
-
-<div class="row">
-    <div class="col-md-12">
-        <table class="table">
-            <tr>
-                <th>
-                    Name
-                </th>
-                <th>
-                    Assignee
-                </th>
-                <th>
-                    Status
-                </th>
-                <th>
-                    Start
-                </th>
-                <th>
-                    End
-                </th>
-            </tr>
-
-            @foreach (var item in Model.OrderBy(t => t.StartDateTime).ThenBy(t => t.Name))
-            {
-                <tr>
-                    <td>
-                        <a asp-controller="Task" asp-action="Details" asp-area="Admin" asp-route-id="@item.Id">@item.Name</a>
-                    </td>
-                    @if (item.AssignedVolunteers == null || item.AssignedVolunteers.Count == 0)
-                    {
-                        <td>*</td>
-                        <td>*</td>
-                    }
-                    else if (item.AssignedVolunteers.Count == 1)
-                    {
-                        <td>@item.AssignedVolunteers[0].UserName</td>
-                        <td>@(item.AssignedVolunteers[0]?.Status ?? "*")</td>
-                    }
-                    else
-                    {
-                        <td><a asp-controller="Task" asp-action="Details" asp-area="Admin" asp-route-eventId="@item.EventId" asp-route-id="@item.Id">Multiple Volunteers</a></td>
-                        <td>@item.AcceptedVolunteerCount / @item.AssignedVolunteers.Count Accepted</td>
-                    }
-
-                    <td>
-                        <time value="item.StartDateTime"></time>                        
-                    </td>
-                    <td>
-                        <time value="item.EndDateTime"></time>
-                    </td>
-                </tr>
-            }
-
-        </table>
-    </div>
-</div>
