+++ AllReadyApp/Web-App/AllReady/Areas/Admin/Views/VolunteerTask/Delete.cshtml
@@ -1,6 +1,5 @@
-@using AllReady.Areas.Admin.ViewModels.Task
-@using System.Linq
-@model DeleteViewModel
+﻿@using System.Linq
+@model AllReady.Areas.Admin.ViewModels.VolunteerTask.DeleteViewModel
 @{  ViewData["Title"] = Model.Title; }
 
 <div class="row">
